fix: remove kakao login page and add kakao login logic into first login screen
