+++ frontend-hybrid/yakal/lib/screens/Login/SocialLogin/screen.dart
@@ -1,27 +1,36 @@
 import 'package:dio/dio.dart';
+import 'package:flutter/foundation.dart' as foundation;
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+import 'package:flutter_svg/flutter_svg.dart';
 import 'package:get/get.dart';
+import 'package:google_sign_in/google_sign_in.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart';
 import 'package:yakal/screens/Login/LoginProcess/login_route.dart';
 import 'package:yakal/screens/Login/LoginProcess/screen.dart';
+import 'package:yakal/utilities/enum/login_platform.dart';
+import 'package:yakal/utilities/style/color_styles.dart';
 import 'package:yakal/viewModels/Profile/user_view_model.dart';
+import 'package:yakal/widgets/Login/social_login_button.dart';
 
-import '../../../utilities/style/color_styles.dart';
+import '../../../widgets/Base/outer_frame.dart';
 
-class KakaoLoginScreen extends StatefulWidget {
-  const KakaoLoginScreen({super.key});
+class SocialLoginScreen extends StatefulWidget {
+  const SocialLoginScreen({super.key});
 
   @override
-  State<KakaoLoginScreen> createState() => _KakaoLoginScreenState();
+  State<SocialLoginScreen> createState() => _SocialLoginScreenState();
 }
 
-class _KakaoLoginScreenState extends State<KakaoLoginScreen> {
+class _SocialLoginScreenState extends State<SocialLoginScreen> {
   final userViewModel = Get.put(UserViewModel(), permanent: true);
   final routeController = Get.put(LoginRouteController());
+
+  bool get isiOS =>
+      foundation.defaultTargetPlatform == foundation.TargetPlatform.iOS;
 
   Future<bool> _kakaoLoginByAccessToken(String accessToken) async {
     var dio = Dio();
@@ -102,8 +111,18 @@ class _KakaoLoginScreenState extends State<KakaoLoginScreen> {
     }
   }
 
-  Future<void> _login() async {
-    var isSuccess = await _kakaoLogin();
+  Future<void> _login(ELoginPlatform loginPlatform) async {
+    late bool isSuccess;
+
+    switch (loginPlatform) {
+      case ELoginPlatform.KAKAO:
+        isSuccess = await _kakaoLogin();
+        break;
+      case ELoginPlatform.GOOGLE:
+        break;
+      case ELoginPlatform.APPLE:
+        break;
+    }
 
     if (isSuccess) {
       if (!context.mounted) {
@@ -113,25 +132,23 @@ class _KakaoLoginScreenState extends State<KakaoLoginScreen> {
       await userViewModel.fetchNameAndMode(context);
 
       routeController.goto(LoginRoute.terms);
-      Get.offNamed("/login/process");
+      Get.toNamed("/login/process");
 
       if (userViewModel.user.value.isAgreedMarketing == null) {
         routeController.goto(LoginRoute.terms);
-        Get.offNamed("/login/process");
+        Get.toNamed("/login/process");
         return;
       }
 
       final regex = RegExp(r'user#\d{6}');
       if (regex.hasMatch(userViewModel.user.value.nickName)) {
         routeController.goto(LoginRoute.nicknameInput);
-        Get.offNamed("/login/process");
+        Get.toNamed("/login/process");
         return;
       }
 
-      Get.offNamed("/");
+      Get.offAllNamed("/");
     } else {
-      Get.back();
-
       if (!context.mounted) {
         return;
       }
@@ -146,20 +163,74 @@ class _KakaoLoginScreenState extends State<KakaoLoginScreen> {
     }
   }
 
-  @override
-  void initState() {
-    super.initState();
-    _login();
+  void _showComingSoonSnackBar(BuildContext context) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      const SnackBar(
+        behavior: SnackBarBehavior.floating,
+        content: Text('기능 준비중입니다.'),
+        duration: Duration(milliseconds: 3000),
+      ),
+    );
   }
 
   @override
   Widget build(BuildContext context) {
-    return const Scaffold(
-      backgroundColor: ColorStyles.white,
-      body: Center(
-        child: CircularProgressIndicator(
-          color: ColorStyles.main,
-        ),
+    return OuterFrame(
+      safeAreaColor: ColorStyles.gray1,
+      outOfSafeAreaColor: ColorStyles.gray1,
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          const SizedBox(
+            height: 200,
+          ),
+          SvgPicture.asset(
+            'assets/icons/login_main.svg',
+          ),
+          const Spacer(),
+          Container(
+            color: ColorStyles.gray1,
+            child: Padding(
+              padding: const EdgeInsets.fromLTRB(20, 30, 20, 70),
+              child: Column(
+                children: [
+                  SocialLoginButton(
+                    text: "카카오로 로그인",
+                    color: ColorStyles.black,
+                    backgroundColor: const Color(0xFFFEE500),
+                    iconPath: "assets/icons/kakao.svg",
+                    onPressed: () {
+                      _login(ELoginPlatform.KAKAO);
+                    },
+                  ),
+                  const SizedBox(
+                    height: 20,
+                  ),
+                  isiOS
+                      ? SocialLoginButton(
+                          text: "Apple로 로그인",
+                          color: ColorStyles.white,
+                          backgroundColor: ColorStyles.black,
+                          iconPath: "assets/icons/apple.svg",
+                          onPressed: () {
+                            _showComingSoonSnackBar(context);
+                          },
+                        )
+                      : SocialLoginButton(
+                          text: "Google 계정으로 로그인",
+                          color: ColorStyles.black,
+                          backgroundColor: ColorStyles.white,
+                          iconPath: "assets/icons/google.svg",
+                          onPressed: () async {
+                            final GoogleSignInAccount? googleUser =
+                                await GoogleSignIn().signIn();
+                          },
+                        ),
+                ],
+              ),
+            ),
+          ),
+        ],
       ),
     );
   }
