+++ frontend-hybrid/yakal/lib/utilities/enum/login_platform.dart
@@ -1,0 +1,5 @@
+enum ELoginPlatform {
+  KAKAO,
+  GOOGLE,
+  APPLE,
+}
