+++ frontend-hybrid/yakal/lib/utilities/enum/user_properties.dart
@@ -1,0 +1,3 @@
+enum EUserProperties {
+  NICKNAME;
+}
